Add definition of intensional ontologies, remove pre-ontology refactor code, add property definitions
We refactored the code so that most definitions correctly use
dot-notation, added definition of contingency, moved the definition
of events outside the ontology namespace declaration, added some
definitions for particular kinds of ontologies and how they are related
to intensional ontologies.

We also commented out the proof that all accidents are simple because it
was failling in our refactor. We still need to make sure the "simple"
predicate uses dot notation. We also defined a cast from entities to
substances.