Closed all sorrys
Commented the first which was an useless one.
Had to add a few definitions concerning intrinsic
composites to close all goals. The original intention
of proving that all composites in potency are
composite in act turned out to be impossible,
so we proved a weaker lemma for intrinsic composites
instead.