Changed "property" to "predicate". Added analogy of being and related definitions.
We reserved the name "property" for a new, more general, experimental definition, defined
in the same module. We also made this module import the state module now, instead of the substances
module, which is now imported indirectly. Also removed universe variables and added the
"noncomputable theory" declaration. To solve a potential philosophical objection, we changed
property.essential declarations to predicate.dere declarations (as in "de re necessary").
property.accidental stayed predicate.accidental for now, but we might consider changing it in the future
to avoid the same objection, and added a TODO note about it.

We also added an experimental definition of "being" and platonic extensions as structures and related definitions.
Based on this definition we will be able to provide a formalization of the 4th way.

We also added many other definitions and a few lemmas. This is a major commit with
a lot of changes.