Proved 2 lemmas and moved causal definitions.
Proved lemmas theism_of_ctheism and simple_of_purely_actual.
Moved cause.first_cause and cause.first_cause to the causality
module, and changed cause.principled definition to a definition
exported by this module (it became an alias). Added
divine_simplicity as a theorem to be proven, and a full
new lemma called ctheism_of_no_universe₂. Made other minor
corrections.

Also commented the god.free theorem declaration because it was
not compiling, as the interface for cause.is_free has changed.
We will correct this in the future.