Made many corrections to causality.lean for thesis
Made minor corrections to multiple parts of causality.lean to
make definitions reasonable to publish in our thesis. The main
additions to the module were a major overhaul of Pruss's argument,
together with the treatment of an objection to it, and the
addition of the Gale-Pruss argument. In ontology.lean we added
some new definitions to support the formalization of the response
to the objection, and we also added an unrelated note about
positions of realism.