Developed similarity.lean with resemblance nominalism
Made many new additions to similarity.lean which are best
explained by the diff. The module is now more readable
and defends a more robust, albeit tentative, idea of essence.
It is compatible both with trope essentialism and
resemblance nominalism. Small modifications to alexandroff_space.lean
were to support this development.